Add level 3 assets
